Explore the effects of the parameters "theta" and "perplexity" of a t-stochastic neighbour embedding or "n_neighors" of a UMAP embedding.

Creates a page with two tabs and a sidebar. The first tab generates the dimension reduction plot of the UMAP or t-SNE with the parameters set in the sidebar by clicking the "Generate plot" button. It is possible to save the scatterplot with its parameters to compare it with another parameters by clicking the button "Add plot for comparison". The second tab "Compare selected plots" contains a grid layout with the saved reduced dimensions. A list in the sidebar shows the saved plots and enables the deletion of plots.

add_dimred_comparison_page(dashboard, object, ...)

# S4 method for i2dashboard,missing
add_dimred_comparison_page(
  dashboard,
  exprs_values,
  seed = NULL,
  calculateUMAP = list(),
  calculateTSNE = list(),
  page = "dimred_comparison_page",
  title = "Explore UMAP/t-SNE parameter",
  menu = "Tools"
)

# S4 method for i2dashboard,SingleCellExperiment
add_dimred_comparison_page(
  dashboard,
  object,
  exprs_values = "logcounts",
  subset_row = NULL,
  ...
)

# S4 method for i2dashboard,SummarizedExperiment
add_dimred_comparison_page(
  dashboard,
  object,
  exprs_values = "logcounts",
  subset_row = NULL,
  ...
)

# S4 method for i2dashboard,Seurat
add_dimred_comparison_page(
  dashboard,
  object,
  assay = "RNA",
  assay_slot = "data",
  subset_row = NULL,
  ...
)

Arguments

dashboard
A i2dashboard.

object
An object of class Seurat, SingleCellExperiment or SummarizedExperiment.

...
Further parameters passed to the core function.

exprs_values
A numeric matrix of log-expression values where rows are features and columns are cells. Alternatively, a character indicating which assay of object to use.

seed
An integer vector, containing the random number generator (RNG) state for random number generation with set.seed().

calculateUMAP
A list with parameters passed on to scater::calculateUMAP()

calculateTSNE
A list with parameters passed on to scater::calculateTSNE()

page
The name of the page to be added.

title
The title of the page.

menu
The menu tab to which this page is subordinated. Default "Tools"

subset_row
An optional character vector (of feature names), a logical vector or numeric vector (of indices) specifying the features to use. The default of NULL will use all features.

assay
A character specifying the assay (object@assays) to obtain expression values from. (Default: "RNA")

assay_slot
A character specifying the name of the data slot in the assay. (Default: "data")

Value

An object of class i2dashboard.