Add a dimension reduction page with feature metadata

This function adds a page with two linked components to the dashboard object: A scatterplot, showing samples in along two-dimensional coordinates, and a table, showing feature metadata. A click on a feature in the table updates the scatterplot with the feature expression.

add_dimred_feature_page(dashboard, object, ...)

# S4 method for i2dashboard,missing
add_dimred_feature_page(
  dashboard,
  use_dimred,
  exprs_values,
  feature_metadata,
  page = "dimred_feature_page",
  title = "Dim. reduction & feature metadata",
  menu = NULL
)

# S4 method for i2dashboard,SingleCellExperiment
add_dimred_feature_page(
  dashboard,
  object,
  use_dimred,
  exprs_values,
  feature_metadata,
  subset_row = NULL,
  ...
)

# S4 method for i2dashboard,Seurat
add_dimred_feature_page(
  dashboard,
  object,
  use_dimred,
  exprs_values,
  feature_metadata,
  subset_row = NULL,
  assay = "RNA",
  assay_slot = "data",
  ...
)

Arguments

dashboard
An object of class i2dashboard.

object
A SingleCellExperiment object or a Seurat object.

...
Further parameters passed to the core function.

use_dimred
Coordinates of the reduced dimensions, used for the scatterplot (see Details).

exprs_values
Expression data of features of interest in rows and samples in columns (see Details).

feature_metadata
A data.frame (matrix) along rows of exprs_values containing feature metadata, or a character vector indicating columns from rowData(object) or object[[assay]]@feature.data.

page
The name of the page to be added.

title
The title of the page.

menu
(Optional) The name of the menu, under which the page should appear.

subset_row
An optional character vector (of feature names), a logical vector or numeric vector (of indices) specifying the features to use. The default of NULL will use all features.

assay
A character specifying the assay (object@assays) to obtain expression values from.

assay_slot
A character specifying the name of the data slot in the assay.

Value

An object of class i2dashboard.

Details

The parameters use_dimred, exprs_values (or assay) and group_by take different arguments depending on the class of object. In case no object is supplied (i2dashboard,missing-method), the parameters are expected to be of class data.frame or matrix. In case a SingleCellExperiment or Seurat object is supplied, the parameters are expected to be of class character, containing

the name of an item in reducedDims(object) or object@reductions,

a valid assay name from assayNames(object) or names(object@assays),

column names of colData(object) or object@meta.data.