Create expression visualization for multiple selected features on a regular grid.

Users can select features and a dimension reduction to plot feature expression values.

add_feature_grid_page(dashboard, object, ...)

# S4 method for i2dashboard,missing
add_feature_grid_page(
  dashboard,
  use_dimred,
  exprs_values,
  title = "Feature grid",
  page = "feature_grid_page",
  menu = "Tools"
)

# S4 method for i2dashboard,SingleCellExperiment
add_feature_grid_page(
  dashboard,
  object,
  use_dimred,
  exprs_values,
  subset_row = NULL,
  ...
)

# S4 method for i2dashboard,Seurat
add_feature_grid_page(
  dashboard,
  object,
  use_dimred,
  assay,
  slot = "data",
  subset_row = NULL,
  ...
)

Arguments

dashboard
An object of class i2dashboard.

object
A SingleCellExperiment object or a Seurat object.

...
Further parameters passed to the core function.

use_dimred
Coordinates of the reduced dimensions, used for the scatterplot (see Details).

exprs_values
Expression data of features of interest in rows and samples in columns (see Details).

title
The title of the page.

page
The name of the page to be added.

menu
(Optional) The name of the menu, under which the page should appear.

subset_row
A character vector (of feature names), a logical vector or numeric vector (of indices) specifying the features to use. The default of NULL will use all features.

assay
A character vector specifying which assay from object@assays to obtain expression values from.

slot
A character vector specifying the name of the slot in the assay.

Value

An object of class i2dashboard.

Details

The parameters use_dimred, exprs_values (or assay) and group_by take different arguments depending on the class of object. In case no object is supplied (i2dashboard,missing-method), the parameters are expected to be of class data.frame or matrix. In case a SingleCellExperiment or Seurat object is supplied, the parameters are expected to be of class character, containing

the name item(s) in reducedDims(object) or object@reductions,

a valid assay name from assayNames(object) or names(object@assays),

column names of colData(object) or object@meta.data.