Renders a component containing a scatterplot with optional selection options

Renders a component containing a scatterplot with optional selection options

scatterplot(dashboard, object, ...)

# S4 method for i2dashboard,missing
scatterplot(
  dashboard,
  x,
  y,
  colour_by = NULL,
  labels = NULL,
  exprs_values = NULL,
  title = NULL,
  x_title = NULL,
  y_title = NULL,
  plot_title = NULL,
  source = "A",
  transmitter = NULL
)

# S4 method for i2dashboard,SingleCellExperiment
scatterplot(
  dashboard,
  object,
  from = c("colData", "rowData", "reducedDim"),
  x = NULL,
  y = NULL,
  colour_by = NULL,
  use_dimred = NULL,
  exprs_values = NULL,
  subset_row = NULL,
  ...
)

# S4 method for i2dashboard,Seurat
scatterplot(
  dashboard,
  object,
  from = c("meta.data", "meta.features", "embedding"),
  x = NULL,
  y = NULL,
  colour_by = NULL,
  use_dimred = NULL,
  assay = "RNA",
  slot = NULL,
  subset_row = NULL,
  ...
)

Arguments

dashboard
An object of class i2dashboard.

object
An object of class Seurat or SingleCellExperiment.

...
Further parameters passed to the core function.

x
Data that will be mapped to the x-axis (see Details).

y
Data that will be mapped to the y-axis (see Details).

colour_by
Numeric or factorial values that will be used for colouring.

labels
An optional vector with sample names. A dropdown menu for colouring by label will be provided.

exprs_values
Expression data of features of interest in rows and samples in columns (see Details).

title
The title of the components junk.

x_title
An optional title of the x-axis. If not provided the column names from x are used instead.

y_title
An optional title of the y-axis. If not provided the column names from y are used instead.

plot_title
An optional title of the plot.

source
A character defining the id used in plotly's source argument of the component.

transmitter
A character defining the id of an existing transmitter to obtain the data from.

from
A character specifying where to obtain the data from object (see Details).

use_dimred
A character vector indicating the reduced dimension to use from "object" (see Details).

subset_row
An optional character vector (of feature names), a logical vector or numeric vector (of indices) specifying the features to use. The default of NULL will use all features.

assay
A character defining the assay of object and is used for obtaining the exprs_values (default "RNA") (see Details).

slot
A character defining the data slot of assay.

Value

A string containing markdown code for the rendered component

Details

The parameters x, y, colour_by, use_dimred, exprs_values, assay and slot) take different arguments depending on the class of object. In case the i2dashboard,missing-method, the parameters x, y, colour_by and exprs_values are expected to be of class data.frame or matrix. The parameters x, y can also be numeric vectors. The parameters from, use_dimred, assay and slot can be ignored. In case the i2dashboard,SingleCellExperiment-method, the parameters are expected to be of class character:

the parameter from can be either "colData", "rowData" or "reducedDim"

use_dimred the name of an item in reducedDims(object)

exprs_values a valid assay name from assayNames(object)

In case of the i2dashboard,Seurat-method, the parameters are expected to be of class character:

the parameter from can be either "meta.data" for sample metadata, "meta.feature" for feature metadata, "embedding" for a dimension reduction

reduction the name of an item in object@reductions

assay a valid assay name from names(object@assays)

slot a valid data slot from assay

In both cases, x, y, colour_by take column names of from.