Renders a component containing an HTML table of summarized sample or feature metadata of single-cell experiment objects or data.frames.

Renders a component containing an HTML table of summarized sample or feature metadata of single-cell experiment objects or data.frames.

summarize_metadata(dashboard, object, ...)

# S4 method for i2dashboard,missing
summarize_metadata(
  dashboard,
  data,
  columns = colnames(data),
  FUNS = c("min", "max", "mean", "median"),
  group_by = NULL,
  caption = NULL,
  description = NULL,
  title = NULL
)

# S4 method for i2dashboard,SingleCellExperiment
summarize_metadata(
  dashboard,
  object,
  from = c("colData", "rowData"),
  columns = NULL,
  group_by = NULL,
  caption = "Table: ",
  description = NULL,
  ...
)

# S4 method for i2dashboard,Seurat
summarize_metadata(
  dashboard,
  object,
  from = c("meta.data", "meta.features"),
  assay = "RNA",
  columns = NULL,
  group_by = NULL,
  caption = "Table",
  description = NULL,
  ...
)

Arguments

dashboard
An object of class i2dashboard.

object
An object of class Seurat or SingleCellExperiment.

...
Further parameters passed to the core function.

data
Data.frame containing numeric observations to summarize.

columns
The metadata columns to summarize.

FUNS
A named vector indicating summary functions.

group_by
A column name of a column containing grouping factors for grouping.

caption
The tables caption.

description
The tables description.

title
The title of the component.

from
A character specifying where to obtain the data from object, e.g. colData/rowData.

assay
A character defining the assay of object.

Value

A string containing markdown code for the rendered component