Renders a component containing a vertical violin plot

Renders a component containing a vertical violin plot

violinplot(dashboard, object, ...)

# S4 method for i2dashboard,missing
violinplot(
  dashboard,
  y,
  group_by = NULL,
  title = NULL,
  y_title = NULL,
  group_by_title = NULL,
  transmitter = NULL
)

# S4 method for i2dashboard,SingleCellExperiment
violinplot(
  dashboard,
  object,
  from = c("colData", "rowData"),
  y = NULL,
  group_by = NULL,
  ...
)

# S4 method for i2dashboard,Seurat
violinplot(
  dashboard,
  object,
  from = c("meta.data", "meta.features"),
  assay = "RNA",
  y = NULL,
  group_by = NULL,
  ...
)

Arguments

dashboard
An object of class i2dashboard.

object
An object of class Seurat or SingleCellExperiment.

...
Further parameters passed to the core function.

y
Data containing numeric observations for the vertical axis (see Details).

group_by
Optionally provide data containing grouping factors for the vertical axis.

title
The title of the component.

y_title
The title of the y-axis.

group_by_title
The title of the x-axis.

transmitter
A character defining the id of an existing transmitter to obtain the data from.

from
A character specifying where to obtain the data from (see Details).

assay
Necessery, if use = "meta.feature". A character defining the assay to obtain the feature metadata from (default "RNA").

Value

A string containing markdown code for the rendered component

Details

The parameters y, group_by take different arguments depending on the class of object. In case of the i2dashboard,missing-method, the parameters are expected to be of class data.frame or matrix. In case of the i2dashboard,SingleCellExperiment- or i2dashboard,Seurat-method, the parameters are expected to be of class character and

the parameter from can be either "colData" or "rowData" for a SingleCellExperiment object or

"meta.data" or "meta.features" for a Seurat object.

In both cases, y and group_by take column names of from.